backend/pyproject.toml
[project]
name = "campaign-os-backend"
version = "0.1.0"
description = "Campaign Operations OS - Backend API"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",

    # Redis
    "redis>=5.0.1",

    # Authentication
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.2",

    # Configuration
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # S3/MinIO
    "boto3>=1.34.0",

    # Utilities
    "python-slugify>=8.0.1",
    "python-dateutil>=2.8.2",

    # Google Calendar integration
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",

    # Encryption for OAuth tokens
    "cryptography>=41.0.0",

    # HTTP client for OAuth token revocation
    "httpx>=0.26.0",

    # Background task scheduling
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
